Trying out
I am trying to find a way to change the pivot of the obcject so that  i can work better (If i use only top point of the rectangle then i would use coordinates better).
I did try to only instantine if in cameraframe, it worked but because i want it there to be an effect of rising in clouds from top, i need to figure out the spawn points again.